feat : TDD 연습을 위한 초기화
