--- src/main/java/lotto/model/Lottos.java
@@ -1,32 +1,0 @@
-package lotto.model;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Lottos {
-    private List<Lotto> lottos;
-
-    public Lottos(List<Lotto> lottos) {
-        this.lottos = lottos;
-    }
-
-    public static Lottos generateLottos(int ticketAmount) {
-        List<Lotto>  arr = new ArrayList<>();
-        for(int i = 0 ;i<ticketAmount ; i++){
-            arr.add(Lotto.generateLottoByRandom());
-        }
-        return new Lottos(List.copyOf(arr));
-    }
-
-    public List<LottoResult> matchLottos(Lotto answer, int bonusNumber) {
-        List<LottoResult> arr = new ArrayList<>();
-        for(Lotto userLotto : lottos) {
-            arr.add(userLotto.matchUp(answer,bonusNumber));
-        }
-        return arr;
-    }
-
-    public List<Lotto> getLottosForMessage() {
-        return lottos;
-    }
-}
